Fix: FKの型変更

+++ kpt/database/migrations/2021_01_25_114010_create_projects_table.php
@@ -22,7 +22,7 @@ class CreateProjectsTable extends Migration
 
             //所有チームID
             //teamsテーブルの外部キー
-            $table->integer('team_id');
+            $table->unsignedBigInteger('team_id');
             $table->foreign('team_id')->references('id')->on('teams');
 
             //作成日時、更新日時
