+++ kpt/database/migrations/2021_01_25_114103_create_goals_table.php
@@ -22,7 +22,7 @@ class CreateGoalsTable extends Migration
 
             //プロジェクトID
             //teamsテーブルの外部キー
-            $table->integer('project_id');
+            $table->unsignedBigInteger('project_id');
             $table->foreign('project_id')->references('id')->on('projects');
 
             //期限
